Разбивает js код на модули
